dashboard/app.py
import streamlit as st
import pandas as pd
import numpy as np
import os
import plotly.express as px
from dotenv import load_dotenv
from supabase import create_client, Client

# ==========================================
# CONFIGURATION & INITIALIZATION
# ==========================================
load_dotenv()
SUPABASE_URL = os.environ.get("SUPABASE_URL", "http://placeholder")
SUPABASE_KEY = os.environ.get("SUPABASE_KEY", "placeholder")

@st.cache_resource
def init_connection() -> Client:
    try:
        return create_client(SUPABASE_URL, SUPABASE_KEY)
    except Exception:
        return None

supabase = init_connection()

st.set_page_config(page_title="AAPT Dashboard", layout="wide", page_icon="🏫")

# Custom App Styling
st.markdown("""
<style>
    .metric-container {
        border-radius: 10px;
        padding: 20px;
        background-color: #f0f2f6;
        margin-bottom: 20px;
        text-align: center;
    }
    .metric-value {
        font-size: 2.5rem;
        font-weight: 700;
        color: #1f77b4;
    }
</style>
""", unsafe_allow_html=True)

# ==========================================
# UI ROUTING AND LAYOUT
# ==========================================
st.sidebar.title("AAPT Controls")
role = st.sidebar.radio("Dashboard View", ["Admin", "Teacher", "Student/Parent"])

st.sidebar.markdown("---")
st.sidebar.info("AAPT uses Advanced Vision pipelines to autonomously track and record student attendance. Powered by Supabase & OpenCV.")

# ==========================================
# ADMIN DASHBOARD
# ==========================================
if role == "Admin":
    st.title("Admin Dashboard ⚙️")
    st.markdown("Manage system rules, hardware configurations, and student enrollment database.")
    
    col1, col2 = st.columns(2)
    with col1:
        st.subheader("System Tuning")
        st.write("Configure dynamic scanning intervals.")
        with st.form("settings_form"):
            scan_interval = st.slider("Timed Scans Interval (minutes)", min_value=1, max_value=60, value=10)
            zoom_enabled = st.checkbox("Enable Digital Quadrant Zooming (High-Fi Detection)", value=True)
            deep_sort = st.checkbox("Enable DeepSORT Tracking", value=True)
            
            if st.form_submit_button("Update System Settings"):
                if supabase:
                    # Mocking an update call for now as we haven't hardcoded an admin UUID
                    st.success("Changes pushed to Production System Settings in Supabase.")
                else:
                    st.success("Changes saved locally for simulation.")
                    
    with col2:
        st.subheader("Facial Data & Enrollment")
        st.write("CRUD Interface for \"Student-by-Student\" Facial Data")
        with st.form("enroll_form"):
            student_name = st.text_input("Full Name")
            student_role = st.selectbox("Role", ["Student", "Teacher"])
            uploaded_file = st.file_uploader("Upload Identity Photo (Clear Face)", type=['png', 'jpg', 'jpeg'])
            
            if st.form_submit_button("Extract & Enroll"):
                if uploaded_file and student_name:
                    st.info(f"Extracting 512D Embeddings from image for {student_name}...")
                    st.success("Successfully enrolled and updated pgvector in Supabase.")
                else:
                    st.warning("Please fill all fields and upload a valid image.")
                    
    st.markdown("---")
    st.subheader("Global Security & Health Logs")
    if supabase:
        try:
            logs = supabase.table("scan_logs").select("*").limit(5).execute()
            if logs.data:
                st.dataframe(pd.DataFrame(logs.data), use_container_width=True)
            else:
                st.write("No database logs generated yet.")
        except Exception:
            st.warning("Database disconnected.")
    else:
        # Mock database logs
        st.dataframe(pd.DataFrame({
            "session_id": ["uuid-1","uuid-2"], "scan_timestamp": ["2026-03-09 10:00:00", "2026-03-09 10:10:00"], 
            "student_id": ["std-1","std-2"], "is_present": [True, False]
        }), use_container_width=True)

# ==========================================
# TEACHER DASHBOARD
# ==========================================
elif role == "Teacher":
    st.title("Teacher Dashboard 👨‍🏫")
    st.markdown("Live classroom metrics and automated disciplinary alerts.")
    
    st.info("🟢 Active Session Detected. Subject: **Computer Science 101** | Time Remaining: **45 mins**")
    
    # KPIs
    c1, c2, c3 = st.columns(3)
    c1.markdown("<div class='metric-container'><h3>Total Enrolled</h3><p class='metric-value'>30</p></div>", unsafe_allow_html=True)
    c2.markdown("<div class='metric-container'><h3>Live Occupancy</h3><p class='metric-value'>28</p></div>", unsafe_allow_html=True)
    c3.markdown("<div class='metric-container'><h3>Bunks / Exits</h3><p class='metric-value' style='color:#d62728'>2</p></div>", unsafe_allow_html=True)

    st.subheader("Disciplinary Alerts (2-Scan Rule)")
    st.error("The following students have missed 2 consecutive autonomous 10-minute interval scans.")
    alerts_data = pd.DataFrame([
        {"Student Name": "John Doe", "Status": "Bunked (Missing for 20m)", "Estimated Time of Flight": "10:20 AM"},
        {"Student Name": "Jane Smith", "Status": "Early Exit", "Estimated Time of Flight": "11:00 AM"}
    ])
    st.table(alerts_data)
    
    st.subheader("Session Roster")
    st.write("Current detected entities in the room.")
    
# ==========================================
# STUDENT/PARENT DASHBOARD
# ==========================================
elif role == "Student/Parent":
    st.title("Student & Parent Portal 🎓")
    st.markdown("Track absolute persistence and overall health of class attendance.")
    
    st.subheader("Personal Presence-over-Time Analytics")
    # Generate mock 30-day analytics
    np.random.seed(42)
    dates = pd.date_range(end=pd.Timestamp.today(), periods=30)
    attendance = [1 if np.random.rand() > 0.15 else 0 for _ in range(30)] 
    df = pd.DataFrame({"Date": dates, "Present": attendance})
    df['Status'] = df['Present'].apply(lambda x: "Present" if x else "Absent (Bunked/Late)")
    
    fig = px.bar(df, x="Date", y="Present", color="Status", 
                 color_discrete_map={"Present": "#2ca02c", "Absent (Bunked/Late)": "#d62728"},
                 title="30-Day Attendance Overview")
    fig.update_layout(yaxis=dict(visible=False))
    st.plotly_chart(fig, use_container_width=True)
    
    st.subheader("Recent Micro-Scans Tracker")
    st.write("Logs from 10-minute facial recognition sweeps.")
    status_data = pd.DataFrame([
        {"Scan Timestamp": "Today - 10:20 AM", "Detection Result": "Present", "Remarks": "Validated via Zone B Zoom"},
        {"Scan Timestamp": "Today - 10:10 AM", "Detection Result": "Present", "Remarks": "Validated via Main Frame"},
        {"Scan Timestamp": "Yesterday - 11:40 AM", "Detection Result": "Missed", "Remarks": "Triggered 1/2 of Bunk Count"},
    ])
    st.table(status_data)

st.markdown("---")
st.caption("Developed for AAPT - Advanced Autonomous Persistence Tracking")
